fix(mobile): address code review blocking items
- Move AgentProfileMetadata from client hook to types/agent.ts (server-safe)
- Remove typescript.ignoreBuildErrors from mobile next.config.ts
- Add outcome validation ('YES'|'NO') to /api/onchain route
- Fix unregister-device docstring to match actual all-devices behavior
- Document why https://localhost is in PRODUCTION_ORIGINS (Capacitor Android WebView)
