+++ apps/web/src/middleware.ts
@@ -16,9 +16,12 @@ const PRODUCTION_ORIGINS = [
   'https://staging.babylon.market',
   'https://app.staging.babylon.market',
   'https://play.staging.babylon.market',
-  // Capacitor mobile app origins
-  'capacitor://localhost', // iOS
-  'https://localhost', // Android
+  // Capacitor mobile app origins.
+  // Capacitor iOS sets Origin: capacitor://localhost; Android sets Origin: https://localhost.
+  // These are the WebView origins for the native shell — not reachable from external browsers.
+  // Attack surface is limited: an attacker would need to already control the device.
+  'capacitor://localhost', // iOS Capacitor WebView
+  'https://localhost', // Android Capacitor WebView
 ] as const;
 
 /**
